+++ NEWS.md
@@ -1,4 +1,4 @@
-# version 1.5-27
+# version 1.5-29
 
 ## bug fixes